DPL: add DataTakingSupport plugin
This is to avoid linking directly against InfoLogger.
