+++ Framework/Core/include/Framework/CommonServices.h
@@ -53,8 +53,6 @@ struct CommonServices {
   static ServiceSpec driverClientSpec();
   static ServiceSpec monitoringSpec();
   static ServiceSpec datatakingContextSpec();
-  static ServiceSpec infologgerContextSpec();
-  static ServiceSpec infologgerSpec();
   static ServiceSpec configurationSpec();
   static ServiceSpec controlSpec();
   static ServiceSpec rootFileSpec();
