+++ Utilities/Mergers/CMakeLists.txt
@@ -14,7 +14,7 @@
 o2_add_library(Mergers
                SOURCES src/MergerAlgorithm.cxx src/IntegratingMerger.cxx src/MergerInfrastructureBuilder.cxx
                        src/MergerBuilder.cxx src/FullHistoryMerger.cxx src/ObjectStore.cxx
-               PUBLIC_LINK_LIBRARIES O2::Framework)
+               PUBLIC_LINK_LIBRARIES O2::Framework AliceO2::InfoLogger)
 
 o2_target_root_dictionary(
   Mergers
